requirements.txt
annotated-types==0.7.0
anyio==4.9.0
audioop-lts==0.2.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cohere==5.14.0
decorator==5.2.1
fastavro==1.10.0
filelock==3.18.0
Flask==3.0.2
Flask-Cors==4.0.0
fsspec==2025.3.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-texttospeech==2.25.1
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
moviepy==2.1.2
numpy==2.2.4
packaging==24.2
pillow==10.4.0
proglog==0.1.10
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
python-dotenv==1.0.0
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
sniffio==1.3.1
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.0.20250306
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
yt-dlp==2025.3.21
